+++ examples/SwdDebugger/fp-lib-table
@@ -1,4 +1,3 @@
 (fp_lib_table
-  (lib (name calisco)(type KiCad)(uri ${KIPRJMOD}/../../../kicad-libs/calisco.pretty)(options "")(descr ""))
-  (lib (name Sparkfun_RF)(type KiCad)(uri ${KIPRJMOD}/../../../kicad-libs/Sparkfun/RF.pretty)(options "")(descr ""))
+  (lib (name edg)(type KiCad)(uri ${KIPRJMOD}/../edg.pretty)(options "")(descr ""))
 )
